chore: add error logging to all resolvers

+++ wedding-api/src/graphql/resolvers/Event.js
@@ -55,8 +55,9 @@ const createEvent = async (parent, { input }, { db }) => {
       },
     };
   } catch (error) {
+    console.error('createEvent', error);
     if (session) await session.abortTransaction();
-    await deleteCognitoUser({ userId });
+    if (userId) await deleteCognitoUser({ userId });
 
     return {
       success: false,
